Fix photo.giveObjHist error on photo change history listing
This was caused by non-existing region reference in photos_history
collection. The patch defaults deleted region item to empty object when
it is retrieved from db using fillRegionsHash.

Photo history listing on client skips it for simplicity (parent or child
regions will be shown as normal). In case if top level region is
deleted, it is shown in history layout as if region was not set
intially.

Fixes #359