Remove babel plugins loose settings.
Both plugins are part of @babel/preset-env and it does not seem possible
to refer to them as if they are installed separately. Node 5.3.0 did not
complain after plugins were removed in a9f263 in favour of
@babel/preset-env, but I suspect those config statements never worked
since that.

`loose` param indicates those plugins to perform loose equality checks
with null instead of strict equality checks against both null and
undefined.

The code does not have a single occurance of
`plugin-proposal-optional-chaining` described notnation, so it as safe
to remove.

The code has a single occurance of
`plugin-proposal-nullish-coalescing-operator` described notation in
profile.changeRestrictions() method, which functionality does not seem
affected by `loose` setting.

As side note, when `loose` is required, Babel suggests to use
`assumptions.noDocumentAll` config setting instead.