Fix querying objects close to north and south poles.
The north/south poles have no representation on a cylindrical map projection used in Leaflet
(EPSG:3857 "WGS 84 Web-Mercator") which is bounded by bbox [[-180, -85.06], [180, 85.06]]
However, we have photos and clusters located beyond those latitudes (as we store them in
EPSG:4326 coordinate reference in Mongo). This patch expands latitudes to be able to query
those objects when top/bottom edge of map is shown on the screen.

Displaying those object on the map works as normal, Leaflet adds those
point to the top/bottom edge of the map.

Fixes second part of #396.