Restrict photo.giveNearestPhotos distance
Since nearSpehre is costly on big distances, especially in low points
density areas, it is reasonable to make it 10km by default, It is used
in Nearest Photos feed in photo view which does not require very distant
photo to show for the price of slow Mongo response.  Method was also in
API use, we restrict max distance to 1000km for the same reason.