Fix RangeError on photos map load.
The issue is caused by asyncronous nature of yandex and google maps
loading. At the time when navigation bar is created, layer might be not
loaded and maxZoom for the map is not set, thus map returns  Infinity.

Ideally some control over sequence is needed, that may be addressed
during refactoring, for now adding a quick fix.